pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["database", "dashboard"]

[project]
name = "lced-care2beauty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "matplotlib>=3.10.1",
    "mysql-connector>=2.2.9",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "streamlit>=1.43.2",
    "streamlit-theme>=0.58.0",
]
